move torch loader from utils.File to TorchFile
